Rewrite helper not to crash on writing on a closed connection
If the defer statement that closes the write end of the in-memory connection is executed while the copy goroutine is still running, this will cause a panic. But since we closed the connection ourselves this error is expected. Error happening after the function argument returns must be ignored as they bear no meaning for the test.