Make ParallelSniffer resilient against data losses
We must detect errors that may cause data from the socket to be lost, otherwise it is not possible to rebuild the original data stream from the client. For example, if the copy returns an error which is not a timeout, it might have faulted while writing to a byte buffer (unlikely, but maybe possible depending on the implementation).

In case of timeouts, we assume that there are no lost data.